Make SudokuSolver folder and move relevant files to this folder
